STAGE/PART_4/1 - POC Design Airbus/requirements.txt
dlib==19.22.99
imutils==0.5.4
numpy==1.24.3
opencv_python==4.7.0.72
PyQt5==5.15.9
PyQt5_sip==12.12.1
pyserial==3.5
